demonstrate that embedded JS bokeh objects cannot use python callbacks
- this commit uses bokeh.embed.components to return a script and div
representation of the latent heat quadmeshes, with the sliders to
adjust time and (for other variables) vertical level.  This does not
work because the dynamic plots depend on a python engine to handle
callbacks, but the (static) generated JS does not have a python
engine.

- This proves, I think, that I really do need to run a bokeh *server*
on a worker dyno.  The server could probably run on the web dyno,
because heroku supports multiple processes, but this really kind of
abuses the spirit of the free resource.